Remove flawed file comparison test
I used `filecmp.cmp` to compare the **dakota.dat** file output from a
test with a known version of the file. The test fails because the
amount of whitespace in the lines of the file can vary, though I
haven't determined if this is by OS, or by Dakota version. I should
write a better test, that compares the values in the file. I haven't
done so yet. In the meantime, I commented out the problematic lines.