Check that the default multidim parameter study uses bounds
This is a somewhat awkward issue in Dakota: the multidim parameter
study *method* must use *variables* with defined `upper_bounds` and
`lower_bounds`; therefore, information must be communicated across
control blocks. I've resolved this by applying bounds if they're not
set, but only in the Experiment, not in the MultidimParameterStudy
class.