Make 'probability_levels' a default keyword
I added `probability_levels` as a default keyword to all UQ methods.
Even though the docs show it as optional, the PolynomialChaos and
StochasticCollocation BMI classes were failing if it (or `response_levels`)
wasn't set. There appears to be no downside to this (famous last words),
since it can be overriden by the user, and it provides helpful stats
in the **dakota.out** file.

Updated unit tests.

This fixes #43.