Converting yaml.dump to unicode
After changing import of StringIO to be from io (for python 3
compatibility) an TypeError appeared (TypeError: initial_value must be
unicode or None, not str).

To fix this, I’ve explicitly converted the output of yaml.dump to
unicode.

This resolves all issues for passing tests on python 2 on my computer.
Python 3 issues still remain.